No issue: Revert multiple commits for Issue #1327.
This was causing merge conflicts and it's better to do the other way
around.

This reverts commit 06fa426eacb55934d46da49e54f8fb1f85d5fb27.
This reverts commit 32f3faefdb210dc5002d7c19c8ee24bbc4ab1fd6.
This reverts commit f9c99f5470745428b5678217cad2d9f8d369d63c.
This reverts commit e6966ed76ea314034dd77de4bc502b72d9313502.
This reverts commit 4774865633eb2cb70c78d8ee8f16028c90ba3460.
