+++ app/src/androidTest/java/org/mozilla/tv/firefox/browser/BrowserBFT.java
@@ -3,16 +3,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.ui;
+package org.mozilla.tv.firefox.browser;
 
 import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.ViewInteraction;
 import android.support.test.espresso.matcher.ViewMatchers;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.uiautomator.UiDevice;
-import android.support.test.uiautomator.UiObject;
-import android.support.test.uiautomator.UiSelector;
 
 import org.junit.After;
 import org.junit.Rule;
@@ -20,26 +17,28 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mozilla.tv.firefox.MainActivity;
 import org.mozilla.tv.firefox.R;
-import org.mozilla.tv.firefox.helpers.SkipOnboardingMainActivityTestRule;
+import org.mozilla.tv.firefox.SkipOnboardingMainActivityTestRule;
 
 import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.pressImeActionButton;
 import static android.support.test.espresso.action.ViewActions.replaceText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.RootMatchers.withDecorView;
 import static android.support.test.espresso.matcher.ViewMatchers.hasFocus;
+import static android.support.test.espresso.matcher.ViewMatchers.isChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.isNotChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static android.support.test.espresso.web.assertion.WebViewAssertions.webMatches;
+import static android.support.test.espresso.web.model.Atoms.getCurrentUrl;
+import static android.support.test.espresso.web.sugar.Web.onWebView;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.core.IsNot.not;
-import static org.junit.Assert.assertFalse;
+import static org.hamcrest.Matchers.containsString;
 
 @RunWith(AndroidJUnit4.class)
-public class PinnedTileTests {
+public class BrowserBFT {
 
     private UiDevice mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
 
@@ -52,7 +51,7 @@ public class PinnedTileTests {
     }
 
     @Test
-    public void testCustomPinnedTile() {
+    public void testBrowserBFT() {
         onView(allOf(withId(R.id.navUrlInput), isDisplayed(), hasFocus()))
                 .perform(replaceText("example.com"))
                 .perform(pressImeActionButton());
@@ -62,29 +61,60 @@ public class PinnedTileTests {
 
         mDevice.pressMenu();
 
-        final ViewInteraction pinButton = onView(ViewMatchers.withId(R.id.pinButton))
+        /* Defaults for a website that is non-tile */
+
+        onView(ViewMatchers.withId(R.id.turboButton))
+                .check(matches(isChecked()));
+
+        onView(ViewMatchers.withId(R.id.pinButton))
                 .check(matches(isNotChecked()));
 
-        pinButton.perform(click());
+        /* Settings */
 
-        onView(withText(R.string.notification_pinned_site))
-                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
+        onView(ViewMatchers.withId(R.id.navButtonSettings))
+                .check(matches(isDisplayed()))
+                .perform(click());
+
+        onView(ViewMatchers.withId(R.id.container))
                 .check(matches(isDisplayed()));
 
-        UiObject newTile = mDevice.findObject(new UiSelector()
-                .resourceId("org.mozilla.tv.firefox.debug:id/tile_title")
-                .text("example")
-                .enabled(true));
+        pressBack();
 
-        newTile.waitForExists(5000);
+        /* Navigation: visit/back/forward */
 
-        pinButton.perform(click());
+        onView(ViewMatchers.withId(R.id.navUrlInput))
+                .check(matches(isDisplayed()))
+                .perform(click())
+                .perform(replaceText("mozilla.org"))
+                .perform(pressImeActionButton());
 
-        onView(withText(R.string.notification_unpinned_site))
-                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
-                .check(matches(isDisplayed()));
+        onWebView()
+                .check(webMatches(getCurrentUrl(), containsString("mozilla.org")));
 
-        newTile.waitUntilGone(5000);
-        assertFalse(newTile.exists());
+        mDevice.pressMenu();
+
+        onView(ViewMatchers.withId(R.id.navUrlInput))
+                .check(matches(withText(containsString("mozilla.org"))));
+
+        onView(ViewMatchers.withId(R.id.pinButton))
+                .check(matches(isNotChecked()));
+
+        onView(ViewMatchers.withId(R.id.navButtonBack))
+                .check(matches(isDisplayed()))
+                .perform(click());
+
+        mDevice.pressMenu();
+
+        onWebView()
+                .check(webMatches(getCurrentUrl(), containsString("example.com")));
+
+        mDevice.pressMenu();
+
+        onView(ViewMatchers.withId(R.id.navButtonForward))
+                .check(matches(isDisplayed()))
+                .perform(click());
+
+        onView(ViewMatchers.withId(R.id.navUrlInput))
+                .check(matches(withText(containsString("mozilla.org"))));
     }
 }
