+++ app/src/androidTest/java/org/mozilla/tv/firefox/home/PinnedTileTests.java
@@ -3,13 +3,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.ui;
+package org.mozilla.tv.firefox.home;
 
 import android.support.test.InstrumentationRegistry;
+import android.support.test.espresso.ViewInteraction;
 import android.support.test.espresso.matcher.ViewMatchers;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.uiautomator.UiDevice;
+import android.support.test.uiautomator.UiObject;
+import android.support.test.uiautomator.UiSelector;
 
 import org.junit.After;
 import org.junit.Rule;
@@ -17,28 +20,26 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mozilla.tv.firefox.MainActivity;
 import org.mozilla.tv.firefox.R;
-import org.mozilla.tv.firefox.helpers.SkipOnboardingMainActivityTestRule;
+import org.mozilla.tv.firefox.SkipOnboardingMainActivityTestRule;
 
 import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.pressImeActionButton;
 import static android.support.test.espresso.action.ViewActions.replaceText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.RootMatchers.withDecorView;
 import static android.support.test.espresso.matcher.ViewMatchers.hasFocus;
-import static android.support.test.espresso.matcher.ViewMatchers.isChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.isNotChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static android.support.test.espresso.web.assertion.WebViewAssertions.webMatches;
-import static android.support.test.espresso.web.model.Atoms.getCurrentUrl;
-import static android.support.test.espresso.web.sugar.Web.onWebView;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.core.IsNot.not;
+import static org.junit.Assert.assertFalse;
 
 @RunWith(AndroidJUnit4.class)
-public class BrowserBFT {
+public class PinnedTileTests {
 
     private UiDevice mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
 
@@ -51,7 +52,7 @@ public class BrowserBFT {
     }
 
     @Test
-    public void testBrowserBFT() {
+    public void testCustomPinnedTile() {
         onView(allOf(withId(R.id.navUrlInput), isDisplayed(), hasFocus()))
                 .perform(replaceText("example.com"))
                 .perform(pressImeActionButton());
@@ -61,60 +62,29 @@ public class BrowserBFT {
 
         mDevice.pressMenu();
 
-        /* Defaults for a website that is non-tile */
-
-        onView(ViewMatchers.withId(R.id.turboButton))
-                .check(matches(isChecked()));
-
-        onView(ViewMatchers.withId(R.id.pinButton))
+        final ViewInteraction pinButton = onView(ViewMatchers.withId(R.id.pinButton))
                 .check(matches(isNotChecked()));
 
-        /* Settings */
+        pinButton.perform(click());
 
-        onView(ViewMatchers.withId(R.id.navButtonSettings))
-                .check(matches(isDisplayed()))
-                .perform(click());
-
-        onView(ViewMatchers.withId(R.id.container))
+        onView(withText(R.string.notification_pinned_site))
+                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
                 .check(matches(isDisplayed()));
 
-        pressBack();
+        UiObject newTile = mDevice.findObject(new UiSelector()
+                .resourceId("org.mozilla.tv.firefox.debug:id/tile_title")
+                .text("example")
+                .enabled(true));
 
-        /* Navigation: visit/back/forward */
+        newTile.waitForExists(5000);
 
-        onView(ViewMatchers.withId(R.id.navUrlInput))
-                .check(matches(isDisplayed()))
-                .perform(click())
-                .perform(replaceText("mozilla.org"))
-                .perform(pressImeActionButton());
+        pinButton.perform(click());
 
-        onWebView()
-                .check(webMatches(getCurrentUrl(), containsString("mozilla.org")));
+        onView(withText(R.string.notification_unpinned_site))
+                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
+                .check(matches(isDisplayed()));
 
-        mDevice.pressMenu();
-
-        onView(ViewMatchers.withId(R.id.navUrlInput))
-                .check(matches(withText(containsString("mozilla.org"))));
-
-        onView(ViewMatchers.withId(R.id.pinButton))
-                .check(matches(isNotChecked()));
-
-        onView(ViewMatchers.withId(R.id.navButtonBack))
-                .check(matches(isDisplayed()))
-                .perform(click());
-
-        mDevice.pressMenu();
-
-        onWebView()
-                .check(webMatches(getCurrentUrl(), containsString("example.com")));
-
-        mDevice.pressMenu();
-
-        onView(ViewMatchers.withId(R.id.navButtonForward))
-                .check(matches(isDisplayed()))
-                .perform(click());
-
-        onView(ViewMatchers.withId(R.id.navUrlInput))
-                .check(matches(withText(containsString("mozilla.org"))));
+        newTile.waitUntilGone(5000);
+        assertFalse(newTile.exists());
     }
 }
