+++ app/src/androidTest/java/org/mozilla/tv/firefox/screenshots/ScreenshotTest.java
@@ -1,4 +1,4 @@
-package org.mozilla.tv.firefox.ui.screenshots;
+package org.mozilla.tv.firefox.screenshots;
 
 import android.app.Instrumentation;
 import android.content.Context;
