+++ app/src/androidTest/java/org/mozilla/tv/firefox/session/ClearSessionTest.kt
@@ -4,7 +4,7 @@
 
 @file:Suppress("TestFunctionName")
 
-package org.mozilla.tv.firefox.ui
+package org.mozilla.tv.firefox.session
 
 import android.support.test.espresso.IdlingRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -13,12 +13,12 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mozilla.tv.firefox.helpers.SkipOnboardingMainActivityTestRule
+import org.mozilla.tv.firefox.SkipOnboardingMainActivityTestRule
 import org.mozilla.tv.firefox.helpers.MockServerHelper
 import org.mozilla.tv.firefox.helpers.SessionLoadedIdlingResource
-import org.mozilla.tv.firefox.ui.robots.browser
-import org.mozilla.tv.firefox.ui.robots.home
-import org.mozilla.tv.firefox.ui.robots.settings
+import org.mozilla.tv.firefox.robots.browser
+import org.mozilla.tv.firefox.robots.home
+import org.mozilla.tv.firefox.robots.settings
 
 @RunWith(AndroidJUnit4::class)
 class ClearSessionTest {
